test & implement: two decimal places on statement amounts

+++ src/main/scala/Ledger.scala
@@ -4,19 +4,19 @@ class Ledger(val transactions: Seq[Transaction] = Seq())
 
 object LedgerTools {
 
-  def deposit(amount: Int, date: LocalDate)(ledger: Ledger = new Ledger()): Ledger = {
+  def deposit(amount: Double, date: LocalDate)(ledger: Ledger = new Ledger()): Ledger = {
     new Ledger(
       ledger.transactions :+ new Transaction(amount, date)
     )
   }
 
-  def withdraw(amount: Int, date: LocalDate)(ledger: Ledger = new Ledger()): Ledger = {
+  def withdraw(amount: Double, date: LocalDate)(ledger: Ledger = new Ledger()): Ledger = {
     new Ledger(
       ledger.transactions :+ new Transaction(-amount, date)
     )
   }
 
-  def balance(ledger: Ledger): Int = {
-    ledger.transactions.foldLeft(0)((runningBalance: Int, transaction: Transaction) => runningBalance + transaction.amount)
+  def balance(ledger: Ledger): Double = {
+    ledger.transactions.foldLeft(0.0)((runningBalance: Double, transaction: Transaction) => runningBalance + transaction.amount)
   }
 }
