+++ src/main/scala/Transaction.scala
@@ -1,3 +1,3 @@
 import java.time.LocalDate
 
-class Transaction(val amount: Int, val date: LocalDate)
+class Transaction(val amount: Double, val date: LocalDate)
